+++ src/codec.js
@@ -49,7 +49,9 @@ const enc = new Map([
   [opcodes.OpenDoor, encoder.OpenDoor],
   [opcodes.SetPCControl, encoder.SetPCControl],
   [opcodes.SetInterlock, encoder.SetInterlock],
-  [opcodes.ActivateKeypads, encoder.ActivateKeypads]
+  [opcodes.ActivateKeypads, encoder.ActivateKeypads],
+
+  [opcodes.RestoreDefaultParameters, encoder.RestoreDefaultParameters]
 ])
 
 /**
@@ -84,7 +86,8 @@ const dec = new Map([
   [0xac, decoder.RefreshTaskList],
   [0xb0, decoder.GetEvent],
   [0xb2, decoder.SetEventIndex],
-  [0xb4, decoder.GetEventIndex]
+  [0xb4, decoder.GetEventIndex],
+  [0xc8, decoder.RestoreDefaultParameters]
 ])
 
 module.exports = {