+++ src/decoder.js
@@ -582,6 +582,20 @@ module.exports = {
       deviceId: uint32(bytes, 4),
       ok: bool(bytes, 8)
     }
+  },
+
+  /**
+    * Decodes the response to a restore-default-parameters request (function code 0xc8).
+    *
+    * @param {buffer}   bytes      64 byte array
+    *
+    * @param {object}   Decoded restore-default-parameters response object
+    */
+  RestoreDefaultParameters: function (bytes) {
+    return {
+      deviceId: uint32(bytes, 4),
+      ok: bool(bytes, 8)
+    }
   }
 }